- alert -> dialog 교체완료. dialog 파일선택 취소처리

+++ scripts/file.js
@@ -50,17 +50,19 @@ module.exports = {
                                     defaultPath: "C:", // 디폴트 경로
                                     properties: ["openDirectory"] // 저장 경로를 폴더로 변경
                                 }).then((result) => {
-                                    savePath = result.filePaths[0]; // 지정 경로
-                                    // fs.writeFileSync(path + '/result.json', JSON.stringify(obj)); // 해당 경로로 result.json 파일 생성
-                                    // event.sender.send('fileResponse', true)
+                                    if (!result.canceled) {
+                                        savePath = result.filePaths[0]; // 지정 경로
+                                        // fs.writeFileSync(path + '/result.json', JSON.stringify(obj)); // 해당 경로로 result.json 파일 생성
+                                        // event.sender.send('fileResponse', true)
 
-                                    let zip = new AdmZip(); // 새로운 zip 파일 생성
-                                    let folderPath = path.join(__dirname, `../static/files/inspect/${args}`); // 해당년도 파일 경로
-                                    zip.addFile('result.json', Buffer.from(JSON.stringify(obj), 'utf-8')); // result.json을 버퍼로 읽은 후 zip 파일에 저장
-                                    zip.addLocalFolder(folderPath, '/'); // 해당년도 파일을 zip 파일에 저장
-                                    zip.writeZip(`${savePath}/result.zip`, () => { // zip 파일을 선택 된 경로에 result.zip 으로 생성
-                                        event.sender.send('fileResponse', true);
-                                    });
+                                        let zip = new AdmZip(); // 새로운 zip 파일 생성
+                                        let folderPath = path.join(__dirname, `../static/files/inspect/${args}`); // 해당년도 파일 경로
+                                        zip.addFile('result.json', Buffer.from(JSON.stringify(obj), 'utf-8')); // result.json을 버퍼로 읽은 후 zip 파일에 저장
+                                        zip.addLocalFolder(folderPath, '/'); // 해당년도 파일을 zip 파일에 저장
+                                        zip.writeZip(`${savePath}/result.zip`, () => { // zip 파일을 선택 된 경로에 result.zip 으로 생성
+                                            event.sender.send('fileResponse', true);
+                                        });
+                                    }
                                 })
                             }
                         })
@@ -80,35 +82,37 @@ module.exports = {
             properties: ['openFile'],
             title: '파일 업로드'
         }).then(async (result) => {
-            let filePath = result.filePaths[0]; // 선택된 파일 경로
-            let fileName = basename(filePath); // 선택된 파일 이름
-            let savePath = path.join(__dirname, `../static/files/inspect/${args.year}/`); // 저장 경로
-            // 저장 경로에 폴더가 없으면 해당 폴더 생성
-            let staticPath = path.join(__dirname, '../static');
-            let filesPath = path.join(__dirname, '../static/files');
-            let inspectPath = path.join(__dirname, '../static/files/inspect');
-            let yearPath = path.join(__dirname, `../static/files/inspect/${args.year}`);
-            if (!fs.existsSync(staticPath)) {
-                fs.mkdirSync(staticPath)
-            }
-            if (!fs.existsSync(filesPath)) {
-                fs.mkdirSync(filesPath)
-            }
-            if (!fs.existsSync(inspectPath)) {
-                fs.mkdirSync(inspectPath)
-            }
-            if (!fs.existsSync(yearPath)) {
-                fs.mkdirSync(yearPath)
-            }
-            if (!fs.existsSync(savePath + args.selectedSeq)) {
-                fs.mkdirSync(savePath + args.selectedSeq);
-            }
+            if (!result.canceled) {
+                let filePath = result.filePaths[0]; // 선택된 파일 경로
+                let fileName = basename(filePath); // 선택된 파일 이름
+                let savePath = path.join(__dirname, `../static/files/inspect/${args.year}/`); // 저장 경로
+                // 저장 경로에 폴더가 없으면 해당 폴더 생성
+                let staticPath = path.join(__dirname, '../static');
+                let filesPath = path.join(__dirname, '../static/files');
+                let inspectPath = path.join(__dirname, '../static/files/inspect');
+                let yearPath = path.join(__dirname, `../static/files/inspect/${args.year}`);
+                if (!fs.existsSync(staticPath)) {
+                    fs.mkdirSync(staticPath)
+                }
+                if (!fs.existsSync(filesPath)) {
+                    fs.mkdirSync(filesPath)
+                }
+                if (!fs.existsSync(inspectPath)) {
+                    fs.mkdirSync(inspectPath)
+                }
+                if (!fs.existsSync(yearPath)) {
+                    fs.mkdirSync(yearPath)
+                }
+                if (!fs.existsSync(savePath + args.selectedSeq)) {
+                    fs.mkdirSync(savePath + args.selectedSeq);
+                }
 
-            fs.copyFile(filePath, `${savePath}${args.selectedSeq}\\${fileName}`, (err) => {
-                if (err) {
-                    console.log('Inspect file upload error:: ', err.message);
-                } else event.sender.send('inspectSaveFileResponse', fileName);
-            });
+                fs.copyFile(filePath, `${savePath}${args.selectedSeq}\\${fileName}`, (err) => {
+                    if (err) {
+                        console.log('Inspect file upload error:: ', err.message);
+                    } else event.sender.send('inspectSaveFileResponse', fileName);
+                });
+            }
         })
     }),
     /**
@@ -177,28 +181,29 @@ module.exports = {
                 properties: ['openFile'],
                 title: '파일 업로드'
             }).then(async (result) => {
-                let res;
-                let year = new Date().getFullYear(); // 현재년도
-                let savePath = path.join(__dirname, `../static/files/inspect/${year}`); // 해당년도 파일 경로
-                let zip = new AdmZip(result.filePaths[0]); // zip 파일 생성
-                let zipEntries = zip.getEntries(); // zip 파일 컨텐츠
-                zipEntries.forEach((e) => {
-                    // zip 파일 중 result.json 파일을 찾는다.
-                    if (e.entryName === 'result.json') {
-                        res = JSON.parse(e.getData().toString('utf8'));
-                    }
-                })
-                zip.extractAllTo(savePath, true); // zip 파일 전체를 해당 경로로 저장한다. (덮어쓰기)
-                fs.unlinkSync(`${savePath}\\result.json`); // 저장 후 result.json 파일 삭제
+                if (!result.canceled) {
+                    let res;
+                    let year = new Date().getFullYear(); // 현재년도
+                    let savePath = path.join(__dirname, `../static/files/inspect/${year}`); // 해당년도 파일 경로
+                    let zip = new AdmZip(result.filePaths[0]); // zip 파일 생성
+                    let zipEntries = zip.getEntries(); // zip 파일 컨텐츠
+                    zipEntries.forEach((e) => {
+                        // zip 파일 중 result.json 파일을 찾는다.
+                        if (e.entryName === 'result.json') {
+                            res = JSON.parse(e.getData().toString('utf8'));
+                        }
+                    })
+                    zip.extractAllTo(savePath, true); // zip 파일 전체를 해당 경로로 저장한다. (덮어쓰기)
+                    fs.unlinkSync(`${savePath}\\result.json`); // 저장 후 result.json 파일 삭제
 
-                let dbPath = path.join(__dirname, '../db');
-                let dbFilePath = path.join(dbPath, '/evaluation.db');
-                db = new sqlite3.Database(dbFilePath);
-                db.serialize(() => {
-                    db.run(`DELETE FROM company`);
-                    db.run(`DELETE FROM admin`);
-                    db.run(`DELETE FROM questions`);
-                    db.run(`
+                    let dbPath = path.join(__dirname, '../db');
+                    let dbFilePath = path.join(dbPath, '/evaluation.db');
+                    db = new sqlite3.Database(dbFilePath);
+                    db.serialize(() => {
+                        db.run(`DELETE FROM company`);
+                        db.run(`DELETE FROM admin`);
+                        db.run(`DELETE FROM questions`);
+                        db.run(`
                         INSERT INTO company(id, code, name, type, address, activity_value, training_max, training_value, protect_max, protect_value, appeal_value, completeYn, year)
                         VALUES(
                         1,
@@ -215,14 +220,14 @@ module.exports = {
                         '${res.company.completeYn}', 
                         '${res.company.year}')
                     `);
-                    res.admin.forEach((e) => {
-                        db.run(`
+                        res.admin.forEach((e) => {
+                            db.run(`
                             INSERT INTO admin
                             VALUES(${e.id}, 1, 1, '${e.name}', '${e.roles}', '${e.email}', '${e.tel}', '${e.phone}', '${e.type}')
                         `)
-                    });
-                    res.questions.forEach((e) => {
-                        db.run(`
+                        });
+                        res.questions.forEach((e) => {
+                            db.run(`
                             INSERT INTO questions
                             VALUES(
                             '${e.id}',
@@ -250,41 +255,42 @@ module.exports = {
                             '${e.self_memo}',
                             '${e.inspect_memo}')
                         `);
-                    });
-                    db.run(`
+                        });
+                        db.run(`
                         UPDATE basic_info
                         SET company_seq = 1
                     `, () => {
-                        // 저장 경로에 폴더가 없으면 해당 폴더 생성
-                        let filePath = path.join(__dirname, '../db/evaluation.db');
-                        let staticPath = path.join(__dirname, '../static');
-                        let filesPath = path.join(__dirname, '../static/files');
-                        let resultPath = path.join(__dirname, '../static/files/result');
-                        let savePath = path.join(__dirname, '../static/files/result/'); // 저장 경로
+                            // 저장 경로에 폴더가 없으면 해당 폴더 생성
+                            let filePath = path.join(__dirname, '../db/evaluation.db');
+                            let staticPath = path.join(__dirname, '../static');
+                            let filesPath = path.join(__dirname, '../static/files');
+                            let resultPath = path.join(__dirname, '../static/files/result');
+                            let savePath = path.join(__dirname, '../static/files/result/'); // 저장 경로
 
-                        if (!fs.existsSync(staticPath)) {
-                            fs.mkdirSync(staticPath)
-                        }
-                        if (!fs.existsSync(filesPath)) {
-                            fs.mkdirSync(filesPath)
-                        }
-                        if (!fs.existsSync(resultPath)) {
-                            fs.mkdirSync(resultPath)
-                        }
-                        if (!fs.existsSync(savePath + year)) {
-                            fs.mkdirSync(savePath + year);
-                        }
-
-                        // /static/files/result/해당년도에 최종 결과 파일 복사
-                        fs.copyFile(filePath, `${savePath}${year}\\evaluation.db`, (err) => {
-                            if (err) {
-                                console.log('Inspect file upload error:: ', err.message);
+                            if (!fs.existsSync(staticPath)) {
+                                fs.mkdirSync(staticPath)
                             }
-                        });
+                            if (!fs.existsSync(filesPath)) {
+                                fs.mkdirSync(filesPath)
+                            }
+                            if (!fs.existsSync(resultPath)) {
+                                fs.mkdirSync(resultPath)
+                            }
+                            if (!fs.existsSync(savePath + year)) {
+                                fs.mkdirSync(savePath + year);
+                            }
 
-                        event.sender.send('getFinalFileResponse', true);
+                            // /static/files/result/해당년도에 최종 결과 파일 복사
+                            fs.copyFile(filePath, `${savePath}${year}\\evaluation.db`, (err) => {
+                                if (err) {
+                                    console.log('Inspect file upload error:: ', err.message);
+                                }
+                            });
+
+                            event.sender.send('getFinalFileResponse', true);
+                        })
                     })
-                })
+                }
             });
         } catch (e) {
             event.sender.send('getFinalFileResponse', false);
@@ -296,14 +302,16 @@ module.exports = {
             defaultPath: "C:", // 디폴트 경로
             properties: ["openDirectory"] // 저장 경로를 폴더로 변경
         }).then((result) => {
-            let savePath = result.filePaths[0]; // 지정 경로
-            let zip = new AdmZip(); // 새로운 zip 파일 생성
-            let folderPath = path.join(__dirname, '../');
-            zip.addLocalFolder(folderPath, '/'); // 해당년도 파일을 zip 파일에 저장
-            zip.writeZip(`${savePath}\\back_up.zip`, () => { // zip 파일을 선택 된 경로에 back_up.zip 으로 생성
-                console.log('writeZip:: ', `${savePath}\\back_up.zip`)
-                event.sender.send('backUpResponse', true);
-            });
+            if (!result.canceled) {
+                let savePath = result.filePaths[0]; // 지정 경로
+                let zip = new AdmZip(); // 새로운 zip 파일 생성
+                let folderPath = path.join(__dirname, '../');
+                zip.addLocalFolder(folderPath, '/'); // 해당년도 파일을 zip 파일에 저장
+                zip.writeZip(`${savePath}\\back_up.zip`, () => { // zip 파일을 선택 된 경로에 back_up.zip 으로 생성
+                    console.log('writeZip:: ', `${savePath}\\back_up.zip`)
+                    event.sender.send('backUpResponse', true);
+                });
+            }
         })
     }),
     restore: ipcMain.on('restore', (event, args) => {
@@ -312,17 +320,19 @@ module.exports = {
             defaultPath: "C:", // 디폴트 경로
             properties: ["openFile"] // 저장 경로를 폴더로 변경
         }).then((result) => {
-            let tmpFolderPath = path.join(__dirname, '..', '../tmp');
-            let zip = new AdmZip(result.filePaths[0]);
+            if (!result.canceled) {
+                let tmpFolderPath = path.join(__dirname, '..', '../tmp');
+                let zip = new AdmZip(result.filePaths[0]);
 
-            // 백업 파일 압축 해제
-            zip.extractAllToAsync(tmpFolderPath, true, true, () => {
-                console.log('extract finish');
-                // restore-event 이벤트 수동으로 발생
-                app.emit('restore-event');
-                event.sender.send('restoreResponse', true);
-                app.quit();
-            });
+                // 백업 파일 압축 해제
+                zip.extractAllToAsync(tmpFolderPath, true, true, () => {
+                    console.log('extract finish');
+                    // restore-event 이벤트 수동으로 발생
+                    app.emit('restore-event');
+                    event.sender.send('restoreResponse', true);
+                    app.quit();
+                });
+            }
         })
     }),
 };
